Move Razor tooling microbenchmarks
