+++ Roslyn.slnx
@@ -423,7 +423,7 @@
   <Folder Name="/Razor/benchmarks/">
     <Project Path="src/Razor/Compiler/Microsoft.AspNetCore.Razor.Microbenchmarks.Compiler/Microsoft.AspNetCore.Razor.Microbenchmarks.Compiler.csproj" />
     <Project Path="src/Razor/Compiler/Microsoft.AspNetCore.Razor.Microbenchmarks.Generator/Microsoft.AspNetCore.Razor.Microbenchmarks.Generator.csproj" />
-    <Project Path="src/Razor/src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Microbenchmarks/Microsoft.AspNetCore.Razor.Microbenchmarks.csproj" />
+    <Project Path="src/Razor/Tooling/Microsoft.AspNetCore.Razor.Microbenchmarks/Microsoft.AspNetCore.Razor.Microbenchmarks.csproj" />
   </Folder>
   <Folder Name="/Razor/build/">
     <File Path="src/Razor/.editorconfig" />
