+++ src/Razor/Tooling/Microsoft.AspNetCore.Razor.Microbenchmarks/.gitignore
