+++ src/Razor/Tooling/Microsoft.AspNetCore.Razor.Microbenchmarks/Extensions.cs
