+++ src/Razor/Tooling/Microsoft.AspNetCore.Razor.Microbenchmarks/Program.cs
