+++ src/Razor/Tooling/Microsoft.AspNetCore.Razor.Microbenchmarks/ResourceSet.cs
