+++ src/Razor/Tooling/Microsoft.AspNetCore.Razor.Microbenchmarks/Resources.cs
