+++ src/Razor/Tooling/Microsoft.AspNetCore.Razor.Microbenchmarks/Resources/Formatting/DocumentFormattingBenchmark.cshtml
