+++ src/Razor/Tooling/Microsoft.AspNetCore.Razor.Microbenchmarks/testapps/ComponentApp/Components/Pages/Weather.razor
