+++ src/Razor/Tooling/Microsoft.AspNetCore.Razor.Microbenchmarks/testapps/ComponentApp/Components/Routes.razor
