+++ src/Razor/Tooling/Microsoft.AspNetCore.Razor.Microbenchmarks/testapps/LargeProject/Views/Home/View001.cshtml
