+++ src/Razor/Tooling/Microsoft.AspNetCore.Razor.Microbenchmarks/testapps/LargeProject/Views/Shared/_ViewStart.cshtml
